Fixed GeneralPath cast error.

+++ src/net/rptools/maptool/client/ui/token/FlowDiamondTokenOverlay.java
@@ -76,11 +76,11 @@ public class FlowDiamondTokenOverlay extends FlowColorDotTokenOverlay {
     protected Shape getShape(Rectangle bounds, Token token) {
         Rectangle2D r = getFlow().getStateBounds2D(bounds, token, getName());
         GeneralPath p = new GeneralPath();
-        p.moveTo(r.getCenterX(), r.getY());
-        p.lineTo(r.getX(), r.getCenterY());
-        p.lineTo(r.getCenterX(), r.getMaxY());
-        p.lineTo(r.getMaxX(), r.getCenterY());
-        p.lineTo(r.getCenterX(), r.getY());
+        p.moveTo((float)r.getCenterX(), (float)r.getY());
+        p.lineTo((float)r.getX(), (float)r.getCenterY());
+        p.lineTo((float)r.getCenterX(), (float)r.getMaxY());
+        p.lineTo((float)r.getMaxX(), (float)r.getCenterY());
+        p.lineTo((float)r.getCenterX(), (float)r.getY());
         p.closePath();
         return p;
     }
